# packages/openapi/schema.yaml
openapi: 3.0.0
info:
  description: API for tracking GitHub commit activity and comparing with rivals
  title: Commitly API
  contact: {}
  version: "1.0"
paths:
  /api/auth/callback:
    post:
      description: フロントエンド（NextAuth.js）からユーザー情報を受け取りDBに保存
      tags:
        - auth
      summary: Github OAuth コールバック
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.CallbackRequest"
        description: コールバックリクエスト
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.UserResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  /api/auth/logout:
    post:
      description: セッションはフロントエンド（NextAuth.js）で管理されるため、バックエンドでは特に処理不要
      tags:
        - auth
      summary: ログアウト
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.MessageResponse"
  /api/dashboard/monthly:
    get:
      security:
        - GitHubUserID: []
      description: 今月のコミット統計を自分とライバルで比較
      tags:
        - dashboard
      summary: 月間ダッシュボードデータを取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/usecase.DashboardData"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  /api/dashboard/weekly:
    get:
      security:
        - GitHubUserID: []
      description: 直近7日間のコミット統計を自分とライバルで比較
      tags:
        - dashboard
      summary: 週間ダッシュボードデータを取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/usecase.DashboardData"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  /api/me:
    get:
      security:
        - GitHubUserID: []
      description: 認証済みユーザーの情報を返す
      tags:
        - user
      summary: 現在のユーザー情報を取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.UserResponse"
  /api/notifications/slack:
    get:
      security:
        - GitHubUserID: []
      description: 現在のSlack通知設定を返す（Webhook URLはマスク済み）
      tags:
        - notifications
      summary: Slack通知設定を取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.SlackNotificationSettingResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
    put:
      security:
        - GitHubUserID: []
      description: Slack通知設定の有効/無効を切り替える
      tags:
        - notifications
      summary: Slack通知の有効/無効を更新
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.UpdateEnabledRequest"
        description: 有効/無効更新リクエスト
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.UpdateEnabledResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
    post:
      security:
        - GitHubUserID: []
      description: Slack Webhook URLを登録して通知設定を作成する
      tags:
        - notifications
      summary: Slack通知設定を作成
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.CreateSlackNotificationRequest"
        description: Slack通知設定作成リクエスト
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.SlackNotificationSettingResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
    delete:
      security:
        - GitHubUserID: []
      description: Slack通知設定を削除する
      tags:
        - notifications
      summary: Slack通知設定を削除
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  /api/rivals:
    get:
      security:
        - GitHubUserID: []
      description: 登録済みのライバル一覧を返す
      tags:
        - rivals
      summary: ライバル一覧を取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.RivalsListResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
    post:
      security:
        - GitHubUserID: []
      description: GitHubユーザー名を指定してライバルを登録する
      tags:
        - rivals
      summary: ライバルを追加
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dto.AddRivalRequest"
        description: ライバル追加リクエスト
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.RivalResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  "/api/rivals/{id}":
    delete:
      security:
        - GitHubUserID: []
      description: 指定IDのライバルを削除する
      tags:
        - rivals
      summary: ライバルを削除
      parameters:
        - description: ライバルID
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/dto.ErrorResponse"
  /health:
    get:
      description: サーバーの稼働状態を確認する
      tags:
        - health
      summary: ヘルスチェック
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dto.HealthResponse"
servers:
  - url: //localhost:8080
components:
  securitySchemes:
    GitHubUserID:
      type: apiKey
      name: X-GitHub-User-ID
      in: header
  schemas:
    dto.AddRivalRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    dto.CallbackRequest:
      type: object
      required:
        - github_user_id
        - github_username
      properties:
        avatar_url:
          type: string
        email:
          type: string
        github_user_id:
          type: integer
        github_username:
          type: string
    dto.CreateSlackNotificationRequest:
      type: object
      properties:
        webhook_url:
          type: string
    dto.ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: エラーメッセージ
    dto.HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    dto.MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 処理が完了しました
    dto.RivalResponse:
      type: object
      required:
        - avatar_url
        - created_at
        - github_user_id
        - github_username
        - id
      properties:
        avatar_url:
          type: string
          example: https://avatars.githubusercontent.com/u/2
        created_at:
          type: string
        github_user_id:
          type: integer
          example: 67890
        github_username:
          type: string
          example: rival-user
        id:
          type: integer
          example: 1
    dto.RivalsListResponse:
      type: object
      required:
        - count
        - max_rivals
        - rivals
      properties:
        count:
          type: integer
          example: 3
        max_rivals:
          type: integer
          example: 5
        rivals:
          type: array
          items:
            $ref: "#/components/schemas/dto.RivalResponse"
    dto.SlackNotificationSettingResponse:
      type: object
      required:
        - created_at
        - id
        - is_enabled
        - updated_at
        - webhook_url
      properties:
        created_at:
          type: string
        id:
          type: integer
          example: 1
        is_enabled:
          type: boolean
          example: true
        updated_at:
          type: string
        webhook_url:
          type: string
          example: https://hooks.slack.com/services/T00...
    dto.UpdateEnabledRequest:
      type: object
      properties:
        is_enabled:
          type: boolean
    dto.UpdateEnabledResponse:
      type: object
      required:
        - is_enabled
      properties:
        is_enabled:
          type: boolean
          example: true
    dto.UserResponse:
      type: object
      required:
        - avatar_url
        - github_user_id
        - github_username
        - id
      properties:
        avatar_url:
          type: string
          example: https://avatars.githubusercontent.com/u/1
        created_at:
          type: string
        github_user_id:
          type: integer
          example: 12345
        github_username:
          type: string
          example: octocat
        id:
          type: integer
          example: 1
    usecase.DailyCommitSummary:
      type: object
      required:
        - commit_count
        - date
      properties:
        commit_count:
          type: integer
        date:
          type: string
    usecase.DashboardData:
      type: object
      required:
        - end_date
        - my_stats
        - period
        - rivals
        - start_date
      properties:
        end_date:
          type: string
        my_stats:
          $ref: "#/components/schemas/usecase.UserCommitStats"
        period:
          description: '"weekly" or "monthly"'
          type: string
        rivals:
          type: array
          items:
            $ref: "#/components/schemas/usecase.UserCommitStats"
        start_date:
          type: string
    usecase.RepositoryCommitSummary:
      type: object
      required:
        - commit_count
        - repository
      properties:
        commit_count:
          type: integer
        repository:
          type: string
    usecase.UserCommitStats:
      type: object
      required:
        - avatar_url
        - daily_stats
        - github_user_id
        - github_username
        - repo_stats
        - total_commits
      properties:
        avatar_url:
          type: string
        daily_stats:
          type: array
          items:
            $ref: "#/components/schemas/usecase.DailyCommitSummary"
        github_user_id:
          type: integer
        github_username:
          type: string
        repo_stats:
          type: array
          items:
            $ref: "#/components/schemas/usecase.RepositoryCommitSummary"
        total_commits:
          type: integer